Add weird identifiers test
